requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
google-cloud-vision>=3.0.0
loguru>=0.7.0
PyMuPDF>=1.23.0
pdf2image>=1.16.0
Pillow>=10.0.0
google-generativeai>=0.3.0
